+++ tests/misc_test.py
@@ -338,3 +338,35 @@ def test_list_checks(
     main(long.list_checks)
     std = capsys.readouterr()
     assert all(i.code in std.out for i in E.values())
+
+
+def test_exclude_argument(
+    init_file: InitFileFixtureType,
+    main: MockMainType,
+) -> None:
+    """Test bash script is ignored when exclude argument passed.
+
+    :param init_file: Initialize a test file.
+    :param main: Mock ``main`` function.
+    """
+    template = """
+#!/usr/bin/env bash
+new-ssl() {
+  domain="${1}"
+  config="${2:-"${domain}.conf"}"
+  openssl \
+    req \
+    -new \
+    -newkey \
+    rsa:2048 \
+    -nodes \
+    -sha256 \
+    -out "${domain}.csr" \
+    -keyout "${domain}.key" \
+    -config "${config}"
+}
+
+new-ssl "${@}"
+"""
+    init_file(template)
+    assert main(".", long.exclude, "file.py") == 0