+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/docsig/compare/v0.44.3...HEAD)
 ------------------------------------------------------------------------
+### Added
+- exclude option
 
 [0.44.3](https://github.com/jshwi/docsig/releases/tag/v0.44.3) - 2024-03-24
 ------------------------------------------------------------------------